[add] change canWord element one to many
